Move training monitor to gui submodule
