+++ sleap/gui/widgets/monitor.py
